DEF-1531: Fixed a couple of issues that was discovered during the code review.

+++ engine/facebook/src/java/com/dynamo/android/facebook/Facebook.java
@@ -159,7 +159,7 @@ public class Facebook implements Handler.Callback {
         startActivity(ACTION_SHOW_DIALOG, extras);
     }
 
-    public void postEvent(String event, long valueToSum, String[] keys, String[] values) {
+    public void postEvent(String event, double valueToSum, String[] keys, String[] values) {
         try {
             if (this.eventLogger != null) {
                 if (keys.length == 0) {
